Refactor the implementation of the app navigation
